% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/professional_survey_features.R
\name{add_participant}
\alias{add_participant}
\title{Add Participant}
\usage{
add_participant(participant)
}
\arguments{
\item{participant}{Participant data}
}
\value{
Participant ID
}
\description{
Add Participant
}
