% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_features.R
\name{add_variable_labels}
\alias{add_variable_labels}
\title{Add Variable Labels}
\usage{
add_variable_labels(data, survey_id)
}
\arguments{
\item{data}{Data frame}

\item{survey_id}{Survey ID}
}
\value{
Data with labels
}
\description{
Add Variable Labels
}
