% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhanced_features.R
\name{auto_save_session}
\alias{auto_save_session}
\title{Auto-Save Session Data}
\usage{
auto_save_session(session, data, force = FALSE)
}
\arguments{
\item{session}{Shiny session object}

\item{data}{Session data to save}

\item{force}{Force immediate save regardless of timer}
}
\value{
Logical indicating success
}
\description{
Automatically saves session data at regular intervals
}
