% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhanced_features.R
\name{batch_process_items}
\alias{batch_process_items}
\title{Batch Process Items}
\usage{
batch_process_items(items, process_fn, batch_size = 100, parallel = FALSE)
}
\arguments{
\item{items}{Items to process}

\item{process_fn}{Processing function}

\item{batch_size}{Batch size}

\item{parallel}{Whether to use parallel processing}
}
\value{
Processed results
}
\description{
Processes items in optimized batches
}
