% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhanced_features.R
\name{block_identifier}
\alias{block_identifier}
\title{Block Identifier}
\usage{
block_identifier(identifier, duration = 3600)
}
\arguments{
\item{identifier}{Identifier to block}

\item{duration}{Block duration in seconds (default: 3600 = 1 hour)}
}
\description{
Blocks an identifier (IP or user) from accessing the system
}
