% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhanced_features.R
\name{cache_result}
\alias{cache_result}
\title{Cache Computation Result}
\usage{
cache_result(key, value, ttl = 3600)
}
\arguments{
\item{key}{Cache key}

\item{value}{Value to cache}

\item{ttl}{Time to live in seconds (default: 3600)}
}
\value{
Logical indicating success
}
\description{
Caches expensive computation results
}
