% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_system.R
\name{calculate_similarity}
\alias{calculate_similarity}
\title{Calculate String Similarity}
\usage{
calculate_similarity(s1, s2)
}
\arguments{
\item{s1}{First string}

\item{s2}{Second string}
}
\value{
Similarity score (0-1)
}
\description{
Calculates similarity between two strings
}
