% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhanced_features.R
\name{check_rate_limit}
\alias{check_rate_limit}
\title{Rate Limiting Check}
\usage{
check_rate_limit(identifier, endpoint, limit = 60)
}
\arguments{
\item{identifier}{User or IP identifier}

\item{endpoint}{API endpoint}

\item{limit}{Maximum requests per minute}
}
\value{
Logical indicating if request is allowed
}
\description{
Checks if request should be rate limited
}
