% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_experience_improvements.R
\name{check_setup}
\alias{check_setup}
\title{Check Setup Completeness}
\usage{
check_setup(config = NULL, item_bank = NULL, verbose = TRUE)
}
\arguments{
\item{config}{Configuration object}

\item{item_bank}{Item bank data frame}

\item{verbose}{Show detailed output}
}
\value{
List of check results with recommendations
}
\description{
Verifies that everything is properly configured
}
