% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhanced_features.R
\name{clean_old_recovery_files}
\alias{clean_old_recovery_files}
\title{Clean Old Recovery Files}
\usage{
clean_old_recovery_files(days)
}
\arguments{
\item{days}{Number of days to retain files}
}
\description{
Removes recovery files older than specified days
}
