% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_error_handling.R
\name{cleanup_old_backups}
\alias{cleanup_old_backups}
\title{Clean Up Old Backups}
\usage{
cleanup_old_backups(keep_count = 5)
}
\arguments{
\item{keep_count}{Number of recent backups to keep}
}
\description{
Removes old backup files, keeping only the most recent ones
}
