% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_session.R
\name{cleanup_old_sessions}
\alias{cleanup_old_sessions}
\title{Clean up old session files to prevent conflicts}
\usage{
cleanup_old_sessions(max_age_hours = 24)
}
\arguments{
\item{max_age_hours}{Maximum age of session files in hours (default: 24)}
}
\description{
Clean up old session files to prevent conflicts
}
