% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_session.R
\name{cleanup_session}
\alias{cleanup_session}
\title{Clean Up Session}
\usage{
cleanup_session(save_final_data = TRUE)
}
\arguments{
\item{save_final_data}{Whether to save final data before cleanup}
}
\description{
Clean Up Session
}
