% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/professional_survey_features.R
\name{compile_logic_rules}
\alias{compile_logic_rules}
\title{Compile Logic Rules}
\usage{
compile_logic_rules(rules)
}
\arguments{
\item{rules}{List of logic rules}
}
\value{
Compiled rules
}
\description{
Compiles branching and skip logic rules
}
