% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_utils.R
\name{core_utils}
\alias{core_utils}
\title{Core Utilities for inrep Package}
\description{
This file consolidates all core utility functions including:
- Basic utility functions (from utils.R)
- Custom operators (from utils_operators.R)  
- Session management utilities (from session_utils.R)
}
\keyword{internal}
