% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhanced_features.R
\name{create_checkbox_group}
\alias{create_checkbox_group}
\title{Create Checkbox Group}
\usage{
create_checkbox_group(input_id, choices, config = list())
}
\arguments{
\item{input_id}{Input ID}

\item{choices}{Choice options}

\item{config}{Configuration}
}
\value{
Shiny UI element
}
\description{
Creates responsive checkbox group
}
