% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_flow_helpers.R
\name{create_default_briefing_content}
\alias{create_default_briefing_content}
\title{Create Default Briefing Content}
\usage{
create_default_briefing_content()
}
\value{
HTML string with default briefing content
}
\description{
Create Default Briefing Content
}
