% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_management.R
\name{create_default_introduction_content}
\alias{create_default_introduction_content}
\title{Create Default Introduction Content}
\usage{
create_default_introduction_content(ui_labels = NULL)
}
\arguments{
\item{ui_labels}{Language labels for UI elements}
}
\value{
HTML string with default introduction content
}
\description{
Create Default Introduction Content
}
