% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_system.R
\name{create_friendly_error}
\alias{create_friendly_error}
\title{Create Friendly Error Message}
\usage{
create_friendly_error(value, valid_values, arg_name, suggestion)
}
\arguments{
\item{value}{Invalid value}

\item{valid_values}{Valid values}

\item{arg_name}{Argument name}

\item{suggestion}{Suggestion text}
}
\value{
Error message
}
\description{
Creates user-friendly error message
}
