% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhanced_features.R
\name{create_likert_scale}
\alias{create_likert_scale}
\title{Create Likert Scale}
\usage{
create_likert_scale(input_id, config = list())
}
\arguments{
\item{input_id}{Input ID}

\item{config}{Configuration}
}
\value{
Shiny UI element
}
\description{
Creates responsive Likert scale
}
