% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhanced_features.R
\name{create_matrix_question}
\alias{create_matrix_question}
\title{Create Matrix Question}
\usage{
create_matrix_question(input_id, options, config = list())
}
\arguments{
\item{input_id}{Input ID}

\item{options}{Matrix options (rows and columns)}

\item{config}{Configuration}
}
\value{
Shiny UI element
}
\description{
Creates responsive matrix/grid question
}
