% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_error_handling.R
\name{create_periodic_backup}
\alias{create_periodic_backup}
\title{Create Periodic Backup}
\usage{
create_periodic_backup(backup_interval = 300)
}
\arguments{
\item{backup_interval}{Backup interval in seconds}
}
\description{
Creates a periodic backup of the current session state
}
