% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhanced_features.R
\name{create_radio_group}
\alias{create_radio_group}
\title{Create Radio Button Group}
\usage{
create_radio_group(input_id, choices, config = list())
}
\arguments{
\item{input_id}{Input ID}

\item{choices}{Choice options}

\item{config}{Configuration}
}
\value{
Shiny UI element
}
\description{
Creates responsive radio button group
}
