% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhanced_features.R
\name{create_ranking_question}
\alias{create_ranking_question}
\title{Create Ranking Question}
\usage{
create_ranking_question(input_id, options, config = list())
}
\arguments{
\item{input_id}{Input ID}

\item{options}{Items to rank}

\item{config}{Configuration}
}
\value{
Shiny UI element
}
\description{
Creates drag-and-drop ranking question
}
