% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_report_generator.R
\name{create_rmd_content}
\alias{create_rmd_content}
\title{Create R Markdown Content}
\usage{
create_rmd_content(study_data, study_config, plot_files, language = "en")
}
\arguments{
\item{study_data}{Study data}

\item{study_config}{Study configuration}

\item{plot_files}{List of plot file paths}

\item{language}{Report language}
}
\value{
R Markdown content as character vector
}
\description{
Create R Markdown Content
}
