% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhanced_features.R
\name{create_slider_input}
\alias{create_slider_input}
\title{Create Slider Input}
\usage{
create_slider_input(input_id, config = list())
}
\arguments{
\item{input_id}{Input ID}

\item{config}{Configuration}
}
\value{
Shiny UI element
}
\description{
Creates responsive slider input
}
