% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhanced_features.R
\name{create_text_input}
\alias{create_text_input}
\title{Create Text Input}
\usage{
create_text_input(input_id, config = list())
}
\arguments{
\item{input_id}{Input ID}

\item{config}{Configuration}
}
\value{
Shiny UI element
}
\description{
Creates responsive text input
}
