% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{create_ui}
\alias{create_ui}
\title{Create Shiny UI}
\usage{
create_ui(config, custom_css = NULL)
}
\arguments{
\item{config}{Study configuration object.}

\item{custom_css}{Optional CSS string.}
}
\value{
Shiny UI object.
}
\description{
Creates the Shiny UI for the adaptive testing application.
}
\keyword{internal}
