% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_start.R
\name{custom_assessment}
\alias{custom_assessment}
\title{Custom Assessment Builder}
\usage{
custom_assessment()
}
\value{
List with config and item_bank
}
\description{
Interactive builder for custom assessments
}
