% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_error_handling.R
\name{emergency_data_preservation}
\alias{emergency_data_preservation}
\title{Emergency Data Preservation}
\usage{
emergency_data_preservation()
}
\value{
Logical indicating if preservation was successful
}
\description{
Attempts to preserve all current data immediately when an error occurs
}
