% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_error_handling.R
\name{emergency_save_current_data}
\alias{emergency_save_current_data}
\title{Emergency Save Current Data}
\usage{
emergency_save_current_data()
}
\description{
Saves current data to a special emergency file
}
