% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_start.R
\name{employee_screening}
\alias{employee_screening}
\title{Employee Screening Template}
\usage{
employee_screening()
}
\value{
List with config and item_bank
}
\description{
Creates an employee assessment with proctoring
}
