% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llm_assistance.R
\name{enable_llm_assistance}
\alias{enable_llm_assistance}
\title{Enable LLM Prompt Generation (Deprecated)}
\usage{
enable_llm_assistance(enable = TRUE, verbose = TRUE)
}
\arguments{
\item{enable}{Logical. TRUE to enable, FALSE to disable.}

\item{verbose}{Logical. Display message.}
}
\value{
Previous setting (invisible).
}
\description{
\strong{Deprecated.} Use
\href{https://selvastics.shinyapps.io/inrep-studio/}{inrep-studio} instead
to configure studies interactively via a GUI.
}
