% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_utils.R
\name{export_data_format}
\alias{export_data_format}
\title{Export Data in Specific Format}
\usage{
export_data_format(data, format, output_dir)
}
\arguments{
\item{data}{Data to export}

\item{format}{Export format}

\item{output_dir}{Output directory}
}
\value{
Export result
}
\description{
Helper function for parallel export.
}
