% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_features.R
\name{export_survey_data}
\alias{export_survey_data}
\title{Data Export Functions}
\usage{
export_survey_data(survey_id, format = "csv", include_labels = TRUE)
}
\arguments{
\item{survey_id}{Survey ID}

\item{format}{Export format}

\item{include_labels}{Include variable labels}
}
\value{
Exported data
}
\description{
Export survey data in various formats
}
