% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_session.R
\name{extend_session}
\alias{extend_session}
\title{Force Session Extension}
\usage{
extend_session(additional_time)
}
\arguments{
\item{additional_time}{Additional time in seconds}
}
\value{
Logical indicating if extension was successful
}
\description{
Force Session Extension
}
