% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_utils.R
\name{fast_item_info}
\alias{fast_item_info}
\title{Fast Item Information Computation with Caching}
\usage{
fast_item_info(theta, item_idx, item_bank, config, cache = NULL)
}
\arguments{
\item{theta}{Ability estimate}

\item{item_idx}{Item index}

\item{item_bank}{Item bank data frame}

\item{config}{Study configuration}

\item{cache}{Cache object for storing results}
}
\value{
Information value
}
\description{
Optimized item information computation with intelligent caching.
}
