% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_system.R
\name{format_values_in_columns}
\alias{format_values_in_columns}
\title{Format Values in Columns}
\usage{
format_values_in_columns(values, columns = 3)
}
\arguments{
\item{values}{Values to format}

\item{columns}{Number of columns}
}
\value{
Formatted string
}
\description{
Formats values in columns for display
}
