% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_management.R
\name{generate_frequency_labels}
\alias{generate_frequency_labels}
\title{Generate frequency scale labels for any number of points}
\usage{
generate_frequency_labels(n_choices, language = "de")
}
\arguments{
\item{n_choices}{Number of response options}

\item{language}{Language code}
}
\value{
Character vector of labels
}
\description{
Generate frequency scale labels for any number of points
}
\keyword{internal}
