% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_report_generator.R
\name{generate_report_plots}
\alias{generate_report_plots}
\title{Generate Plots for PDF Report}
\usage{
generate_report_plots(study_data, study_config, output_dir)
}
\arguments{
\item{study_data}{Study data}

\item{study_config}{Study configuration}

\item{output_dir}{Output directory for plots}
}
\value{
List of plot file paths
}
\description{
Generate Plots for PDF Report
}
