% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhanced_features.R
\name{generate_salt}
\alias{generate_salt}
\title{Generate Salt}
\usage{
generate_salt(length = 32)
}
\arguments{
\item{length}{Length of salt in characters}
}
\value{
Character string with salt
}
\description{
Generates a cryptographically secure salt
}
