% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_session.R
\name{generate_session_id}
\alias{generate_session_id}
\title{Generate Unique Session ID}
\usage{
generate_session_id()
}
\value{
Character string with unique session identifier
}
\description{
Generate Unique Session ID
}
