% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_system.R
\name{generate_suggestion}
\alias{generate_suggestion}
\title{Generate Suggestion}
\usage{
generate_suggestion(value, valid_values, arg_name)
}
\arguments{
\item{value}{Invalid value}

\item{valid_values}{Valid values}

\item{arg_name}{Argument name}
}
\value{
Suggestion string
}
\description{
Generates helpful suggestion for user
}
