% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_error_handling.R
\name{get_current_environment_data}
\alias{get_current_environment_data}
\title{Get Current Environment Data}
\usage{
get_current_environment_data()
}
\value{
List with current environment data
}
\description{
Collects all relevant data from the current environment
}
