% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/language_labels.R
\name{get_language_labels}
\alias{get_language_labels}
\title{Get Language Labels}
\usage{
get_language_labels(language = "en")
}
\arguments{
\item{language}{Language code (en, de, es, fr, fa)}
}
\value{
List of language labels
}
\description{
Retrieves all labels for a specific language
}
