% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_features.R
\name{get_quota_status}
\alias{get_quota_status}
\title{Get Quota Status}
\usage{
get_quota_status(survey_id)
}
\arguments{
\item{survey_id}{Survey ID}
}
\value{
Quota status
}
\description{
Get Quota Status
}
