% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/professional_survey_features.R
\name{get_response_timeline}
\alias{get_response_timeline}
\title{Get Response Timeline}
\usage{
get_response_timeline(responses)
}
\arguments{
\item{responses}{Response list}
}
\value{
Timeline data
}
\description{
Get Response Timeline
}
