% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/professional_survey_features.R
\name{get_survey_responses}
\alias{get_survey_responses}
\title{Get Survey Responses}
\usage{
get_survey_responses(survey_id)
}
\arguments{
\item{survey_id}{Survey ID}
}
\value{
Data frame of responses
}
\description{
Get Survey Responses
}
