% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/professional_survey_features.R
\name{get_survey_statistics}
\alias{get_survey_statistics}
\title{Real-time Statistics Dashboard}
\usage{
get_survey_statistics(survey_id)
}
\arguments{
\item{survey_id}{Survey ID}
}
\value{
Statistics object
}
\description{
Get real-time survey statistics
}
