% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{get_theme_css_path}
\alias{get_theme_css_path}
\title{Get Theme CSS File Path}
\usage{
get_theme_css_path(theme)
}
\arguments{
\item{theme}{Theme name}
}
\value{
Full path to CSS file
}
\description{
Get Theme CSS File Path
}
