% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhanced_features.R
\name{get_theme_styles}
\alias{get_theme_styles}
\title{Get Theme Styles}
\usage{
get_theme_styles(theme_name)
}
\arguments{
\item{theme_name}{Name of theme}
}
\value{
CSS string
}
\description{
Returns theme-specific styles
}
