% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_utils.R
\name{\%||\%}
\alias{\%||\%}
\title{Null-coalescing Operator}
\usage{
x \%||\% y

x \%||\% y
}
\arguments{
\item{lhs}{Left-hand side value}

\item{rhs}{Right-hand side value (default)}
}
\value{
lhs if not NULL, otherwise rhs

lhs if not NULL, otherwise rhs
}
\description{
Returns the left-hand side if not NULL, otherwise returns the right-hand side

Returns the left-hand side if not NULL, otherwise returns the right-hand side
}
\examples{
NULL \%||\% "default"  # Returns "default"
"value" \%||\% "default"  # Returns "value"
NULL \%||\% "default"  # Returns "default"
"value" \%||\% "default"  # Returns "value"
}
