% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_utils.R
\name{\%r\%}
\alias{\%r\%}
\title{String Repetition Operator}
\usage{
string \%r\% times
}
\arguments{
\item{string}{Character string to repeat}

\item{times}{Number of times to repeat}
}
\value{
Character string repeated the specified number of times
}
\description{
Repeats a string a specified number of times
}
\examples{
"=" \%r\% 50  # Returns "=================================================="
"-" \%r\% 10  # Returns "----------"
}
