% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_system.R
\name{group_values_by_category}
\alias{group_values_by_category}
\title{Group Values by Category}
\usage{
group_values_by_category(values)
}
\arguments{
\item{values}{Values to group}
}
\value{
Named list of grouped values
}
\description{
Groups values into logical categories
}
