% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhanced_features.R
\name{handle_branching_rules}
\alias{handle_branching_rules}
\title{Handle Complex Branching Rules}
\usage{
handle_branching_rules(rules)
}
\arguments{
\item{rules}{List of branching rules}
}
\value{
Processed rules
}
\description{
Validates and processes complex branching logic
}
