% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_features.R
\name{initialize_survey_features}
\alias{initialize_survey_features}
\title{Initialize Survey Features}
\usage{
initialize_survey_features(config = list())
}
\arguments{
\item{config}{Survey configuration}
}
\value{
Invisible NULL
}
\description{
Sets up survey capabilities
}
