% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/language_labels.R
\name{is_supported_language}
\alias{is_supported_language}
\title{Validate Language Code}
\usage{
is_supported_language(language)
}
\arguments{
\item{language}{Language code to validate}
}
\value{
Logical indicating if language is supported
}
\description{
Validate Language Code
}
