% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/language_labels.R
\docType{data}
\name{LANGUAGE_DICTIONARY}
\alias{LANGUAGE_DICTIONARY}
\title{Comprehensive Multilingual Support for inrep Package}
\format{
An object of class \code{list} of length 4.
}
\usage{
LANGUAGE_DICTIONARY
}
\description{
Provides complete translations for all UI elements, messages, and content
Supports: English (en), German (de), Spanish (es), French (fr)
}
\keyword{datasets}
