% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{launch_theme_editor}
\alias{launch_theme_editor}
\title{Launch Theme Editor}
\usage{
launch_theme_editor()
}
\value{
Opens the theme editor in the default browser.
}
\description{
Opens a web-based GUI for creating custom themes.
}
\examples{
\dontrun{
# Launch the theme editor
launch_theme_editor()
}
}
