% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/language_labels.R
\name{list_supported_languages}
\alias{list_supported_languages}
\title{List Supported Languages}
\usage{
list_supported_languages()
}
\value{
Vector of supported language codes
}
\description{
List Supported Languages
}
