% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprehensive_data_manager.R
\name{log_action}
\alias{log_action}
\title{Log Action}
\usage{
log_action(action_type, action_details = NULL, page_id = NULL)
}
\arguments{
\item{action_type}{Type of action}

\item{action_details}{Action details}

\item{page_id}{Page identifier}
}
\description{
Logs a user action for comprehensive tracking.
}
