% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_utils.R
\name{log_open}
\alias{log_open}
\title{Log Open}
\usage{
log_open(path)
}
\arguments{
\item{path}{File path for the log.}
}
\value{
TRUE if successful, FALSE otherwise.
}
\description{
Opens a log file connection.
}
