% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprehensive_data_manager.R
\name{log_page_time}
\alias{log_page_time}
\title{Log Page Time}
\usage{
log_page_time(page_id, time_spent_seconds)
}
\arguments{
\item{page_id}{Page identifier}

\item{time_spent_seconds}{Time spent in seconds}
}
\description{
Logs time spent on a page.
}
