% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_features.R
\name{manage_participants}
\alias{manage_participants}
\title{Manage Participants}
\usage{
manage_participants(action, participant = NULL)
}
\arguments{
\item{action}{Action to perform}

\item{participant}{Participant data}
}
\value{
Result of action
}
\description{
Participant management system
}
