% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhanced_features.R
\name{memoize_function}
\alias{memoize_function}
\title{Memoized Function Wrapper}
\usage{
memoize_function(fn, cache_key_fn = NULL)
}
\arguments{
\item{fn}{Function to memoize}

\item{cache_key_fn}{Function to generate cache key from arguments}
}
\value{
Memoized function
}
\description{
Creates a memoized version of a function for performance
}
