% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_utils.R
\name{monitor_parallel_performance}
\alias{monitor_parallel_performance}
\title{Performance Monitoring for Parallel Operations}
\usage{
monitor_parallel_performance(
  operation_name,
  start_time,
  end_time,
  n_items,
  n_workers
)
}
\arguments{
\item{operation_name}{Name of the operation}

\item{start_time}{Start time}

\item{end_time}{End time}

\item{n_items}{Number of items processed}

\item{n_workers}{Number of workers used}
}
\value{
Performance metrics
}
\description{
Monitors performance of parallel operations.
}
