% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_utils.R
\name{null_coalescing_operator}
\alias{null_coalescing_operator}
\alias{\%||\%}
\title{Null-coalescing Operator}
\usage{
x \%||\% y
}
\arguments{
\item{x}{Left-hand side value.}

\item{y}{Right-hand side value (default).}
}
\value{
`x` if not NULL, otherwise `y`.
}
\description{
Returns the left-hand side if not NULL, otherwise returns the right-hand side
}
\examples{
NULL \%||\% "default"  # Returns "default"
"value" \%||\% "default"  # Returns "value"
}
