% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhanced_features.R
\name{optimize_query}
\alias{optimize_query}
\title{Optimize Query}
\usage{
optimize_query(query, data, use_index = TRUE)
}
\arguments{
\item{query}{Query or operation}

\item{data}{Data to query}

\item{use_index}{Whether to use indexing}
}
\value{
Query result
}
\description{
Optimizes and caches database/data queries
}
