% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_utils.R
\name{parallel_batch_process}
\alias{parallel_batch_process}
\title{Parallel Batch Processing for Multiple Participants}
\usage{
parallel_batch_process(participants, item_bank, config, process_fun)
}
\arguments{
\item{participants}{List of participant data}

\item{item_bank}{Item bank data frame}

\item{config}{Study configuration}

\item{process_fun}{Function to process each participant}
}
\value{
List of results for each participant
}
\description{
Processes multiple participants in parallel for batch operations.
}
