% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_utils.R
\name{parallel_export}
\alias{parallel_export}
\title{Parallel Data Export}
\usage{
parallel_export(
  data,
  formats = c("csv", "json", "rds"),
  output_dir = ".",
  config
)
}
\arguments{
\item{data}{Data to export}

\item{formats}{Vector of export formats ("csv", "json", "rds")}

\item{output_dir}{Output directory}

\item{config}{Study configuration}
}
\value{
List of export results
}
\description{
Exports data in multiple formats in parallel.
}
