% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_utils.R
\name{parallel_item_info}
\alias{parallel_item_info}
\title{Parallel Item Information Computation}
\usage{
parallel_item_info(theta, item_indices, item_bank, config, parallel_env)
}
\arguments{
\item{theta}{Ability estimate}

\item{item_indices}{Vector of item indices}

\item{item_bank}{Item bank data frame}

\item{config}{Study configuration}

\item{parallel_env}{Parallel processing environment}
}
\value{
Vector of information values
}
\description{
Computes item information for multiple items in parallel.
}
