% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_features.R
\name{parse_condition}
\alias{parse_condition}
\title{Parse Condition}
\usage{
parse_condition(condition)
}
\arguments{
\item{condition}{Condition string or object}
}
\value{
Parsed condition
}
\description{
Parses logic condition
}
