% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{print.inrep_simulation}
\alias{print.inrep_simulation}
\title{Print an inrep_simulation Object}
\usage{
\method{print}{inrep_simulation}(x, ...)
}
\arguments{
\item{x}{An \code{inrep_simulation} object.}

\item{...}{Ignored.}
}
\description{
Print an inrep_simulation Object
}
