% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_config.R
\name{print.survey_config}
\alias{print.survey_config}
\title{Survey Configuration Methods}
\usage{
\method{print}{survey_config}(x, ...)
}
\arguments{
\item{x}{Survey configuration object}

\item{...}{Additional arguments}
}
\description{
Survey Configuration Methods
}
