% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/professional_survey_features.R
\name{process_conjoint}
\alias{process_conjoint}
\title{Process Conjoint}
\usage{
process_conjoint(q)
}
\arguments{
\item{q}{Question object}
}
\value{
Processed question
}
\description{
Process Conjoint
}
