% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_features.R
\name{process_constant_sum}
\alias{process_constant_sum}
\title{Process Constant Sum}
\usage{
process_constant_sum(q)
}
\arguments{
\item{q}{Question object}
}
\value{
Processed question
}
\description{
Process Constant Sum
}
