% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_features.R
\name{process_file_question}
\alias{process_file_question}
\title{Process File Question}
\usage{
process_file_question(q)
}
\arguments{
\item{q}{Question object}
}
\value{
Processed question
}
\description{
Process File Question
}
