% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/professional_survey_features.R
\name{process_likert_question}
\alias{process_likert_question}
\title{Process Likert Question}
\usage{
process_likert_question(q)
}
\arguments{
\item{q}{Question object}
}
\value{
Processed question
}
\description{
Process Likert Question
}
