% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/professional_survey_features.R
\name{process_matrix_question}
\alias{process_matrix_question}
\title{Process Matrix Question}
\usage{
process_matrix_question(q)
}
\arguments{
\item{q}{Question object}
}
\value{
Processed question
}
\description{
Process Matrix Question
}
