% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_features.R
\name{process_max_diff}
\alias{process_max_diff}
\title{Process Max Diff}
\usage{
process_max_diff(q)
}
\arguments{
\item{q}{Question object}
}
\value{
Processed question
}
\description{
Process Max Diff
}
