% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_features.R
\name{process_multiple_choice}
\alias{process_multiple_choice}
\title{Process Multiple Choice Question}
\usage{
process_multiple_choice(q)
}
\arguments{
\item{q}{Question object}
}
\value{
Processed question
}
\description{
Process Multiple Choice Question
}
