% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_features.R
\name{process_numeric_question}
\alias{process_numeric_question}
\title{Process Numeric Question}
\usage{
process_numeric_question(q)
}
\arguments{
\item{q}{Question object}
}
\value{
Processed question
}
\description{
Process Numeric Question
}
