% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_management.R
\name{process_page_flow}
\alias{process_page_flow}
\title{Process custom page flow for UI rendering}
\usage{
process_page_flow(
  config,
  rv,
  input,
  output,
  session,
  item_bank,
  ui_labels,
  logger,
  auto_close_time = 300,
  auto_close_time_unit = "seconds",
  disable_auto_close = FALSE
)
}
\description{
Process custom page flow for UI rendering
}
