% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_features.R
\name{process_ranking_question}
\alias{process_ranking_question}
\title{Process Ranking Question}
\usage{
process_ranking_question(q)
}
\arguments{
\item{q}{Question object}
}
\value{
Processed question
}
\description{
Process Ranking Question
}
