% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/professional_survey_features.R
\name{process_semantic_diff}
\alias{process_semantic_diff}
\title{Process Semantic Differential}
\usage{
process_semantic_diff(q)
}
\arguments{
\item{q}{Question object}
}
\value{
Processed question
}
\description{
Process Semantic Differential
}
