% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/professional_survey_features.R
\name{process_single_choice}
\alias{process_single_choice}
\title{Process Single Choice Question}
\usage{
process_single_choice(q)
}
\arguments{
\item{q}{Question object}
}
\value{
Processed question
}
\description{
Process Single Choice Question
}
