% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/professional_survey_features.R
\name{process_slider_question}
\alias{process_slider_question}
\title{Process Slider Question}
\usage{
process_slider_question(q)
}
\arguments{
\item{q}{Question object}
}
\value{
Processed question
}
\description{
Process Slider Question
}
