% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_features.R
\name{process_text_question}
\alias{process_text_question}
\title{Process Text Question}
\usage{
process_text_question(q)
}
\arguments{
\item{q}{Question object}
}
\value{
Processed question
}
\description{
Process Text Question
}
