% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhanced_features.R
\name{recover_from_browser_storage}
\alias{recover_from_browser_storage}
\title{Recover Session from Browser Storage}
\usage{
recover_from_browser_storage(session, callback)
}
\arguments{
\item{session}{Shiny session object}

\item{callback}{Function to call with recovered data}
}
\description{
Attempts to recover session data from browser localStorage
}
