% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_management.R
\name{render_demographics_page}
\alias{render_demographics_page}
\title{Render demographics page}
\usage{
render_demographics_page(page, config, rv, ui_labels)
}
\description{
Render demographics page
}
