% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_management.R
\name{render_items_page}
\alias{render_items_page}
\title{Render items page with pagination}
\usage{
render_items_page(page, config, rv, item_bank, ui_labels, session)
}
\description{
Render items page with pagination
}
