% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_management.R
\name{render_results_page}
\alias{render_results_page}
\title{Render results page}
\usage{
render_results_page(
  page,
  config,
  rv,
  item_bank,
  ui_labels,
  auto_close_time = 300,
  auto_close_time_unit = "seconds",
  disable_auto_close = FALSE,
  session = NULL
)
}
\description{
Render results page
}
