% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_error_handling.R
\name{robust_warning_handler}
\alias{robust_warning_handler}
\title{Robust Warning Handler}
\usage{
robust_warning_handler(warning_message)
}
\arguments{
\item{warning_message}{Warning message}
}
\description{
Robust Warning Handler
}
