% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhanced_features.R
\name{sanitize_input}
\alias{sanitize_input}
\title{Sanitize Input for XSS Prevention}
\usage{
sanitize_input(input, allow_html = FALSE)
}
\arguments{
\item{input}{User input to sanitize}

\item{allow_html}{Whether to allow safe HTML tags}
}
\value{
Sanitized input
}
\description{
Sanitizes user input to prevent XSS attacks
}
