% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhanced_features.R
\name{save_to_browser_storage}
\alias{save_to_browser_storage}
\title{Save to Browser Storage}
\usage{
save_to_browser_storage(session, data)
}
\arguments{
\item{session}{Shiny session object}

\item{data}{Data to save}
}
\description{
Saves session data to browser localStorage for refresh recovery
}
