% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhanced_features.R
\name{secure_password}
\alias{secure_password}
\title{Secure Password Storage}
\usage{
secure_password(password, salt = NULL)
}
\arguments{
\item{password}{Plain text password}

\item{salt}{Optional salt for hashing}
}
\value{
List with encrypted password and salt
}
\description{
Securely stores passwords using encryption
}
