% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_start.R
\name{show_examples}
\alias{show_examples}
\title{Show Example Code}
\usage{
show_examples(task = NULL)
}
\arguments{
\item{task}{Task to show example for}
}
\description{
Displays example code for common tasks
}
