% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_management.R
\name{simulate_clinical_studies}
\alias{simulate_clinical_studies}
\title{Simulate Clinical Psychology Studies}
\usage{
simulate_clinical_studies()
}
\value{
List of simulation results
}
\description{
Tests various clinical assessment scenarios
}
