% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_management.R
\name{simulate_corporate_studies}
\alias{simulate_corporate_studies}
\title{Simulate Corporate HR Assessments}
\usage{
simulate_corporate_studies()
}
\value{
List of simulation results
}
\description{
Tests various workplace assessment scenarios
}
