% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_management.R
\name{simulate_educational_studies}
\alias{simulate_educational_studies}
\title{Simulate Educational Assessment Studies}
\usage{
simulate_educational_studies()
}
\value{
List of simulation results with errors and fixes
}
\description{
Tests various educational assessment scenarios that users might create
}
