% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhanced_features.R
\name{start_auto_save_timer}
\alias{start_auto_save_timer}
\title{Start Auto-Save Timer}
\usage{
start_auto_save_timer(interval)
}
\arguments{
\item{interval}{Save interval in seconds}
}
\description{
Starts the automatic save timer
}
