% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_start.R
\name{test_assessment}
\alias{test_assessment}
\title{Test Assessment in Demo Mode}
\usage{
test_assessment(config, item_bank, n_simulated = 5)
}
\arguments{
\item{config}{Configuration object}

\item{item_bank}{Item bank data frame}

\item{n_simulated}{Number of simulated responses}
}
\value{
Test results summary
}
\description{
Runs a quick test of the assessment configuration
}
