% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_management.R
\name{update_item_bank_scale}
\alias{update_item_bank_scale}
\title{Update item bank ResponseCategories for any scale}
\usage{
update_item_bank_scale(item_bank, n_points, start_value = 1)
}
\arguments{
\item{item_bank}{Data frame with item bank}

\item{n_points}{Number of response points}

\item{start_value}{Starting value (default: 1)}
}
\value{
Updated item bank with ResponseCategories column
}
\description{
This function updates all items in an item bank to use the specified
number of response points.
}
\examples{
# Update all items to use 7-point scale
updated_bank <- update_item_bank_scale(bfi_items, 7)

# Update all items to use 10-point scale starting from 0
updated_bank <- update_item_bank_scale(bfi_items, 10, 0)
}
