% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprehensive_data_manager.R
\name{update_page_start_time}
\alias{update_page_start_time}
\title{Update Page Start Time}
\usage{
update_page_start_time(page_id)
}
\arguments{
\item{page_id}{Page identifier}
}
\description{
Updates the start time for the current page.
}
