% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhanced_features.R
\name{validate_action}
\alias{validate_action}
\title{Validate Action}
\usage{
validate_action(action)
}
\arguments{
\item{action}{Action string}
}
\value{
Validated action
}
\description{
Validates branching action
}
