% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_system.R
\name{validate_boolean}
\alias{validate_boolean}
\title{Validate Boolean Parameter}
\usage{
validate_boolean(value, param_name)
}
\arguments{
\item{value}{User-provided value}

\item{param_name}{Parameter name}
}
\value{
Logical value
}
\description{
Validates boolean parameters with fuzzy matching
}
