% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhanced_features.R
\name{validate_condition_expression}
\alias{validate_condition_expression}
\title{Validate Condition Expression}
\usage{
validate_condition_expression(condition)
}
\arguments{
\item{condition}{Condition string}
}
\value{
Validated condition
}
\description{
Validates branching condition expressions
}
