% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/professional_survey_features.R
\name{validate_logic_action}
\alias{validate_logic_action}
\title{Validate Logic Action}
\usage{
validate_logic_action(action)
}
\arguments{
\item{action}{Action object}
}
\value{
Validated action
}
\description{
Validates logic action
}
