% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_system.R
\name{validate_model}
\alias{validate_model}
\title{Validate Model Argument}
\usage{
validate_model(model)
}
\arguments{
\item{model}{Model name provided by user}
}
\value{
Validated model name
}
\description{
Validates and corrects IRT model names
}
