% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_system.R
\name{validate_progress_style}
\alias{validate_progress_style}
\title{Validate Progress Style}
\usage{
validate_progress_style(style)
}
\arguments{
\item{style}{Progress style}
}
\value{
Validated style
}
\description{
Validates progress bar style
}
