% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhanced_features.R
\name{validate_recovery_data}
\alias{validate_recovery_data}
\title{Validate Recovery Data}
\usage{
validate_recovery_data(data)
}
\arguments{
\item{data}{Recovery data to validate}
}
\value{
Logical indicating if data is valid
}
\description{
Validates recovered session data for integrity
}
