% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_config.R
\name{validate_survey_config}
\alias{validate_survey_config}
\title{Validate Survey Configuration}
\usage{
validate_survey_config(config)
}
\arguments{
\item{config}{Survey configuration object}
}
\value{
Logical indicating if configuration is valid
}
\description{
Validate Survey Configuration
}
