% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_system.R
\name{validate_theme}
\alias{validate_theme}
\title{Validate Theme Argument}
\usage{
validate_theme(theme)
}
\arguments{
\item{theme}{Theme name provided by user}
}
\value{
Validated theme name
}
\description{
Validates and corrects theme names
}
