% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhanced_features.R
\name{verify_password}
\alias{verify_password}
\title{Verify Password}
\usage{
verify_password(password, stored_hash, salt)
}
\arguments{
\item{password}{Plain text password to verify}

\item{stored_hash}{Stored password hash}

\item{salt}{Salt used for hashing}
}
\value{
Logical indicating if password is valid
}
\description{
Verifies a password against stored hash
}
