% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhanced_features.R
\name{with_index}
\alias{with_index}
\title{With Index Helper}
\usage{
with_index(data, operation)
}
\arguments{
\item{data}{Data frame}

\item{operation}{Operation to perform}
}
\value{
Operation result
}
\description{
Executes operation with temporary index
}
