% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{zzz}
\alias{zzz}
\title{Package Load/Unload Hooks}
\description{
Handles package loading and unloading to prevent installation conflicts
}
\keyword{Internal}
\keyword{and}
\keyword{globals}
\keyword{imports}
\keyword{internal}
