% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_utils.R
\name{compute_item_info_single}
\alias{compute_item_info_single}
\title{Single Item Information Computation}
\usage{
compute_item_info_single(theta, item_idx, item_bank, config)
}
\arguments{
\item{theta}{Ability estimate}

\item{item_idx}{Item index}

\item{item_bank}{Item bank data frame}

\item{config}{Study configuration}
}
\value{
Information value
}
\description{
Computes information for a single item (used in parallel processing).
}
