% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_flow_helpers.R
\name{create_custom_demographic_ui}
\alias{create_custom_demographic_ui}
\title{Create Custom Demographic UI}
\usage{
create_custom_demographic_ui(demographic_configs, theme = NULL)
}
\arguments{
\item{demographic_configs}{List of demographic configurations}

\item{theme}{Theme configuration}
}
\value{
Shiny UI elements for demographics
}
\description{
Create Custom Demographic UI
}
