% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_management.R
\name{create_demographic_input}
\alias{create_demographic_input}
\title{Create demographic input element}
\usage{
create_demographic_input(
  input_id,
  demo_config,
  input_type,
  current_value = NULL,
  language = "de"
)
}
\description{
Create demographic input element
}
