% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_management.R
\name{create_filter_page}
\alias{create_filter_page}
\title{Create filter page with actual functionality}
\usage{
create_filter_page(input, config)
}
\description{
Create filter page with actual functionality
}
