% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_system.R
\name{generate_theme_css}
\alias{generate_theme_css}
\title{Generate Theme CSS}
\usage{
generate_theme_css(theme_name = "light")
}
\arguments{
\item{theme_name}{Name of the theme}
}
\value{
CSS string
}
\description{
Generates CSS from theme configuration
}
