% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_utils.R
\name{generate_uuid}
\alias{generate_uuid}
\title{Generate UUID}
\usage{
generate_uuid()
}
\value{
Character string containing a UUID.
}
\description{
Generates a universally unique identifier using base R random sampling.
}
