% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_system.R
\name{get_available_themes}
\alias{get_available_themes}
\title{Get Available Themes}
\usage{
get_available_themes()
}
\value{
Character vector of theme names
}
\description{
Lists all available theme files
}
