% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/language_labels.R
\name{get_language_name}
\alias{get_language_name}
\title{Get Language Name}
\usage{
get_language_name(language_code)
}
\arguments{
\item{language_code}{Language code (en, de, es, fr)}
}
\value{
Full language name
}
\description{
Get Language Name
}
