% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llm_assistance.R
\name{get_llm_assistance_settings}
\alias{get_llm_assistance_settings}
\title{Get LLM Assistance Status}
\usage{
get_llm_assistance_settings()
}
\value{
Logical.
}
\description{
Get LLM Assistance Status
}
