% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_system.R
\name{get_theme_css}
\alias{get_theme_css}
\title{Get Theme CSS}
\usage{
get_theme_css(theme = "Light", custom_css = NULL)
}
\arguments{
\item{theme}{Theme name}

\item{custom_css}{Additional custom CSS}
}
\value{
CSS string
}
\description{
Loads theme CSS from file system
}
