% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhanced_features.R
\name{handle_extreme_parameters}
\alias{handle_extreme_parameters}
\title{Handle Extreme Parameters}
\usage{
handle_extreme_parameters(params)
}
\arguments{
\item{params}{List of parameters}
}
\value{
Sanitized parameters
}
\description{
Handles extreme parameter values gracefully
}
