% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_management.R
\name{study_management}
\alias{study_management}
\alias{.inrep_is_debug}
\title{Study Management for inrep Package}
\usage{
.inrep_is_debug()
}
\description{
This file consolidates all study management functions including:
- Study flow helpers (from study_flow_helpers.R)
- Study simulations (from study_simulations.R)
- Custom page flow (from custom_page_flow.R)
- Flow validation (from custom_page_flow_validation.R)
}
\keyword{internal}
