% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_system.R
\docType{data}
\name{theme_system}
\alias{theme_system}
\title{Theme System for increp Package}
\description{
This file consolidates all theme-related functions including:
- Theme definitions (from themes.R)
- CSS loading functions (from get_theme_css.R)
- Font consistency (from ensure_font_consistency.R)
}
\keyword{Definitions}
\keyword{Package}
\keyword{Provides}
\keyword{Theme}
\keyword{UI}
\keyword{comprehensive}
\keyword{definitions}
\keyword{for}
\keyword{inrep}
\keyword{internal}
\keyword{the}
\keyword{theme}
