% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{validate_theme_name}
\alias{validate_theme_name}
\title{Validate Theme Name}
\usage{
validate_theme_name(theme_name)
}
\arguments{
\item{theme_name}{Theme name to validate}
}
\value{
Validated theme name if found, otherwise 'light'
}
\description{
Validate Theme Name
}
